Changed the interpreter to work using pointers
Changed the interpreter to use pointers to values instead of just values.  This will allow the access of the memory by the correct address when iterating in a loop.  Also in the binary expression interpreter we are currently malloc for each binary operation which is slow.  This should be fixed later so that the left value is used instead