Change to values in parser/interpreter
Reverted the other attempt at fixing this for loops.  Now trying to give the name of the variable and then find the variable using its name.  Will need to check for errors.  Also currently in the assignmentexpression function within the interpreter the leftValue of the assignment is assumed to be a value.  Need to do some check on this.